Fix for handling of `coordinates` metadata
The Position TIFF tags (X/Y) were used to populate a missing coordinates tuple for the MIBITIFF image. The coordinates tuple was then used to calculate the (now) missing Position TIFF tags (X/Y) in the MIBITIFF image.
The problem is that these two sets of variables are not necessarily interchangeable: they are equivalent, when both X and Y value are zero or positive. However, the coordinates from the stage can  be negative, while the YPosition TIFF tag must not be negative.

It looks like a regression issue where the initial developer was aware of the TIFF tag specs and introduced the coordinates variable, but then later developers treated these variables interchangeably. Of note is the use of `coordinates` instead of a corresponding position variable in the `tiff.write` function.